refact: refactoring utilities functions
